requirements.txt
requests
beautifulsoup4
PyPDF2
python-docx
openpyxl
xlrd
